build: Formatting: Put ')' on their own lines.
While I'm not very fond of the look of it, it has one clear advantage:
Adding files to the list will make one added line in the diff. When the closing
parenthesis is on the same line as the last item, adding to the end of the list
will cause an additional changed line to be added to the diff.

This style is only applied to file(-like) lists, because they change regularly.