class: Add safe_to, sprikle BOOST_(UN)LIKELY.
Also remove check for lua_type == LUA_TUSERDATA from is_apollo_instance(), for
the reasons indicated in the comment. It still is accountable for over a third
of the runtime of a microbenchmark measuring the duration of a call (from Lua)
to a C++ function taking a dummy object.

The BOOST_LIKELY around is_apollo_instance() really resulted in a nice,
measureable speedup!

The safe_to in general is of course also very valuable.